WIP: Working on playlist information from clients

+++ app/src/main/java/com/example/strawberry_app/data/entity/PlaylistSongEntity.kt
@@ -24,7 +24,7 @@ import androidx.room.Index
     indices = [Index("songId")]
 )
 data class PlaylistSongEntity(
-    val playlistId: Int,
-    val songId: Int,
+    val playlistId: Long,
+    val songId: Long,
     val position: Int
 )
