+++ app/src/main/java/com/example/strawberry_app/screens/ServerValues.kt
@@ -6,10 +6,10 @@ enum class PlayState{
     PLAYING, PAUSED, STOPPED
 }
 data class ServerValues(
-    val activePlaylist: Int = -1,
-    val currentPlaylist: Int = -1,
+    val activePlaylist: Long = -1L,
+    val currentPlaylist: Long = -1L,
     val currentSong: SongInfo = SongInfo(),
-    val currentSongId: Long = -1,
+    val currentSongId: Long = -1L,
     val currentTime: Long = 0L,
     val playState: PlayState = PlayState.STOPPED,
     val volume: Int = 0
